@@ -1,13 +1,13 @@
 plugins {
-	id 'org.springframework.boot' version '2.7.14'
-	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+	id 'org.springframework.boot' version '3.1.2'
+	id 'io.spring.dependency-management' version '1.1.0'
 	id 'java'
-	id "io.freefair.lombok" version "4.1.6"
+	id 'io.freefair.lombok' version '8.2.2'
 }
 
 group = 'com.dd.kanban'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 repositories {
 	mavenCentral()
