@@ -3,7 +3,7 @@ package com.dd.kanban.controller;
 import java.net.URI;
 import java.util.List;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
