@@ -1,6 +1,6 @@
 package com.dd.kanban.request;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 import lombok.Getter;
 
