@@ -10,7 +10,7 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+//import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -27,7 +27,7 @@ import com.dd.kanban.security.JwtAuthenticationFilter;
 		jsr250Enabled = true,
 		prePostEnabled = true
 		)
-public class SecurityConfig extends WebSecurityConfigurerAdapter{
+public class SecurityConfig {
 	@Autowired
 	CustomUserDetailsService customUserDetailsService;
 
@@ -41,48 +41,48 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter{
 
 	}
 
-	@Override
-	public void configure(AuthenticationManagerBuilder authenticationManagerBuilder) throws Exception {
-		authenticationManagerBuilder
-		.userDetailsService(customUserDetailsService)
-		.passwordEncoder(passwordEncoder());
-	}
-
-	@Bean(BeanIds.AUTHENTICATION_MANAGER)
-	@Override
-	public AuthenticationManager authenticationManagerBean() throws Exception {
-		return super.authenticationManagerBean();
-	}
+//	@Override
+//	public void configure(AuthenticationManagerBuilder authenticationManagerBuilder) throws Exception {
+//		authenticationManagerBuilder
+//		.userDetailsService(customUserDetailsService)
+//		.passwordEncoder(passwordEncoder());
+//	}
+//
+//	@Bean(BeanIds.AUTHENTICATION_MANAGER)
+//	@Override
+//	public AuthenticationManager authenticationManagerBean() throws Exception {
+//		return super.authenticationManagerBean();
+//	}
 
 	@Bean
 	public PasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
 	}
 
-	@Override
-	protected void configure(HttpSecurity httpSecurity) throws Exception {
-		httpSecurity
-		.cors()
-		.and()
-		.csrf()
-		.disable()
-		.exceptionHandling()
-		.authenticationEntryPoint(unauthorizedHandler)
-		.and()
-		.sessionManagement()
-		.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-		.and()
-		.authorizeRequests()
-		.antMatchers("/", "/favicon.ico","/**/*.png","/**/*.gif","/**/*.svg","/**/*.jpg","/**/*.html","/**/*.css","/**/*.js")
-		.permitAll()
-		.antMatchers("/api/auth/**")
-		.permitAll()
-		.antMatchers("/api/user/checkUsernameAvailability", "/api/user/checkEmailAvailability")
-		.permitAll()
-		.antMatchers(HttpMethod.GET, "/api/polls/**", "/api/users/**")
-		.permitAll()
-		.anyRequest()
-		.authenticated();
-		httpSecurity.addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
-	}
+//	@Override
+//	protected void configure(HttpSecurity httpSecurity) throws Exception {
+//		httpSecurity
+//		.cors()
+//		.and()
+//		.csrf()
+//		.disable()
+//		.exceptionHandling()
+//		.authenticationEntryPoint(unauthorizedHandler)
+//		.and()
+//		.sessionManagement()
+//		.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+//		.and()
+//		.authorizeRequests()
+//		.antMatchers("/", "/favicon.ico","/**/*.png","/**/*.gif","/**/*.svg","/**/*.jpg","/**/*.html","/**/*.css","/**/*.js")
+//		.permitAll()
+//		.antMatchers("/api/auth/**")
+//		.permitAll()
+//		.antMatchers("/api/user/checkUsernameAvailability", "/api/user/checkEmailAvailability")
+//		.permitAll()
+//		.antMatchers(HttpMethod.GET, "/api/polls/**", "/api/users/**")
+//		.permitAll()
+//		.anyRequest()
+//		.authenticated();
+//		httpSecurity.addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+//	}
 }
